Configure Doorkeeper for correct access by admins
This configures Doorkeeper to only give access to users with the right
level of access.

The resource_owner_authenticator block is checked when Doorkeeper needs
to access a user's data, and returns either an authorized user or
returns headers/redirects as appropriate.

The admin_authenticator block does the same, but returns an authorized
user if they are able to manage the OAuth applications that are allowed
to connect to the site.