Add Doorkeeper profile response
This action enables OAuth clients to load limited profile data (as
defined by the scopes that an application is granted).

The URL this profile is mounted at (`/game_control/profile.json`) needs
to be known by the application, there's no way to discover it through
the OAuth process. Even the JSON format is just a convention.